+++ dap4_test/test_data.sh
@@ -3,7 +3,6 @@
 if test "x$srcdir" = x ; then srcdir=`pwd`; fi
 . ../test_common.sh
 
-set -x
 set -e
 
 . ${top_srcdir}/dap4_test/d4test_common.sh